requirements.txt
# ============================================
# TSUANA - 3D VR World Generator
# Production Dependencies
# ============================================

# Core Dependencies
# -----------------
python-dotenv==1.0.0      # Environment variable management
requests==2.31.0           # HTTP requests for API calls
fastapi==0.115.5           # Web framework for API
uvicorn==0.32.1            # ASGI server for FastAPI

# OpenAI API
# ----------
openai==1.58.1             # AI-powered scene planning

# Scientific Computing
# --------------------
numpy==1.26.4              # Numerical operations for positioning/transforms

# Optional: Advanced 3D Processing
# ---------------------------------
# Uncomment if you need local 3D processing capabilities
# trimesh==4.0.0           # 3D mesh processing
# pygltflib==1.16.1        # GLTF file manipulation
# pillow==10.1.0           # Image processing for textures

# Development Dependencies (Optional)
# -----------------------------------
# pytest==7.4.3            # Testing framework
# black==23.12.0           # Code formatting
# mypy==1.7.1              # Type checking
